(向原) 各機能をコンポネント化
